+++ packages/nostrify/CHANGELOG.md
@@ -5,12 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 0.42.0 - 2025-04-10
 
 ### Added
 
 - Added `@nostrify/react` package.
 - NPool: add `.group` method to create sub-pools.
+
+### Removed
+
+- Removed NSimplePool.
 
 ## 0.41.2 - 2025-04-10